Crate hopr_transport_network

Network and physical transport related utilities

This package contains objects and utilities that allow storing and querying the current state of the observable Network.

Background processes and event updates ensure that the Network data is regularly updated to offer the complete information about each observed peer:

the timestamps of last observations

quality measurements based on the heartbeat and ping reporting mechanism,

individual metrics counting windowed quality characteristics

on and off-chain observed peer related data (multiaddresses…)

NOTE: This object has not been significantly altered from the previous versions and will be improved in the upcoming releases together with physical transport upgrades that will make it possible to record more data faster. As such, this object should slowly transform into a pollable physical network graph processing live telemetry from the packet transport process.

Modules

config

Configuration of the network module.

constants

Global constants published from this crate.

errors

Errors that can be generated by the crate.

heartbeat

Implementation of the main loop for the heartbeat mechanism.

messaging

Low-level transport protocol messaging definitions for ping.

network

Implementation of the peer network logic

ping

Implementation of the pinging mechanism used by the heartbeat and manual pings.

Structs

PeerId

Identifier of a peer of the network.

Traits

HoprDbPeersOperations