hopr_transport_network::ping

Trait PingExternalAPI

pub trait PingExternalAPI {
    // Required method
    fn on_finished_ping<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        peer: &'life1 PeerId,
        result: &'life2 Result<Duration>,
        version: String,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
}

External behavior that will be triggered once a ping operation result is available per each pinged peer.

Required Methods

fn on_finished_ping<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, peer: &'life1 PeerId, result: &'life2 Result<Duration>, version: String, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,